# Performance optimization of RegularGridInterpolator