#7197:
```
RegularGridInterpolator:    CPU:   371.503 us   +/-  9.201 (min:   365.464 / max:   849.020) us     GPU-0:   378.049 us   +/-  9.323 (min:   371.712 / max:   860.160) us
interp(...)         :    CPU: 30634.061 us   +/- 254.224 (min: 30564.859 / max: 32679.887) us     GPU-0: 30642.441 us   +/- 255.291 (min: 30573.120 / max: 32696.831) us
interp._prepare_xi(...):    CPU:  7857.799 us   +/- 28.457 (min:  7842.555 / max:  8298.093) us     GPU-0:  7864.969 us   +/- 29.520 (min:  7849.760 / max:  8322.400) us
interp._find_indices(...):    CPU:  5274.640 us   +/- 17.603 (min:  5252.196 / max:  5564.324) us     GPU-0:  6205.481 us   +/- 17.351 (min:  6183.936 / max:  6491.136) us
interp._evaluate_linear(...):    CPU:  2017.064 us   +/- 171.669 (min:  1982.883 / max:  3876.088) us     GPU-0: 16489.220 us   +/- 12.630 (min: 16456.703 / max: 16524.288) us
interpn(...)        :    CPU: 32392.145 us   +/- 98.787 (min: 32339.826 / max: 33149.544) us     GPU-0: 32400.340 us   +/- 100.352 (min: 32347.904 / max: 33170.208) us
```

This PR:
```
RegularGridInterpolator:    CPU:   310.917 us   +/-  6.714 (min:   305.368 / max:   745.539) us     GPU-0:   316.290 us   +/-  6.795 (min:   311.296 / max:   754.688) us
interp(...)         :    CPU:  3884.846 us   +/- 70.359 (min:  3856.265 / max:  4535.406) us     GPU-0: 22573.148 us   +/- 36.356 (min: 22548.479 / max: 22903.809) us
interp._prepare_xi(...):    CPU:  1782.385 us   +/- 119.381 (min:  1759.524 / max:  4739.485) us     GPU-0:  1788.788 us   +/- 120.395 (min:  1765.568 / max:  4762.624) us
interp._find_indices(...):    CPU:   884.298 us   +/- 31.464 (min:   871.013 / max:  1484.830) us     GPU-0:  5088.810 us   +/- 12.216 (min:  5075.968 / max:  5319.680) us
interp._evaluate_linear(...):    CPU:  1137.647 us   +/- 50.195 (min:  1117.633 / max:  1694.179) us     GPU-0: 15574.334 us   +/- 20.472 (min: 15557.632 / max: 15795.200) us
interpn(...)        :    CPU:  5576.094 us   +/- 205.239 (min:  5518.584 / max:  7364.412) us     GPU-0: 24237.019 us   +/- 35.296 (min: 24208.384 / max: 24537.088) us
```

<details>
benchmark:

```py
import cupy
from cupy import testing
from cupyx.scipy.interpolate import RegularGridInterpolator, interpn
from cupyx.profiler import benchmark


def f(x, y, z):
    return 2 * x**3 + 3 * y**2 - z

x = cupy.linspace(0, 10, 1024)
y = cupy.linspace(0, 10, 1024)
z = cupy.linspace(0, 10, 1024)
xg, yg, zg = cupy.meshgrid(x, y, z, indexing='ij', sparse=True)
data = f(xg, yg, zg)

perf = benchmark(RegularGridInterpolator, ((x, y, z), data), max_duration=2)
print(perf)

interp = RegularGridInterpolator((x, y, z), data)
samples = testing.shaped_random((8, 1024, 1024, 3), xp=cupy, scale=10, dtype='float32')
perf = benchmark(interp, (samples,), name='interp(...)', max_duration=2)
print(perf)

perf = benchmark(interp._prepare_xi, (samples,), name='interp._prepare_xi(...)', max_duration=2)
print(perf)
xi, xi_shape, ndim, nans, out_of_bounds = interp._prepare_xi(samples)

perf = benchmark(interp._find_indices, (xi.T,), name='interp._find_indices(...)', max_duration=2)
print(perf)
indices, norm_distances = interp._find_indices(xi.T)

perf = benchmark(interp._evaluate_linear, (indices, norm_distances), name='interp._evaluate_linear(...)', max_duration=2)
print(perf)

perf = benchmark(interpn, ((x, y, z), data, samples), name='interpn(...)', max_duration=2)
print(perf)
```