@ideasrule 2023-01-05
Two of the changes look great, but I'm sure about this one:

is_nans = cp.isnan(xi).T
nans = is_nans[0].copy()
for is_nan in is_nans[1:]:
     cp.logical_or(nans, is_nan, nans)

It's a lot longer, less elegant, and harder to understand than "nans = cp.any(cp.isnan(xi), axis=-1)".  Does it really make the code that much faster?  If so, we should probably be optimizing cp.any instead.