@takagi 2023-01-11
> Two of the changes look great, but I'm sure about this one:
> 
> ```
> is_nans = cp.isnan(xi).T
> nans = is_nans[0].copy()
> for is_nan in is_nans[1:]:
>     cp.logical_or(nans, is_nan, nans)
> ```
> 
> It's a lot longer, less elegant, and harder to understand than "nans = cp.any(cp.isnan(xi), axis=-1)". Does it really make the code that much faster? If so, we should probably be optimizing cp.any instead.

Sorry for being late. We prefer this implementation compared to the current `cp.any` (or the reduction kernel) because this is more efficient for a shorter series of data. Contrarily, `cp.any` is suitable for longer and more general data. Also, we agree that we should optimize `cp.any`, however, it would take some effort to make it generally work nicer, so let us leave the current `cp.any` as it is here.